Add BZ speed per state and release
Use: --speed -r [RELEASE] -v [VERSION] to see the average days to get to a
certain state. Bugs are separated based on their priority and if they
are RFE or not. Compatible with --squad (Default is OVN). It also prints
a table that outputs the number of bugs that got a certain state to give
context to the data outputed. I put an example of use in the README with
fictional results.